+++ predicators/ground_truth_models/__init__.py
@@ -129,9 +129,7 @@ utils.import_submodules(__path__, __name__)
 
 def deprecated_get_gt_nsrts(env_name: str) -> Set[NSRT]:
     """Create ground truth NSRTs for an env."""
-    if env_name == "stick_button":
-        nsrts = _get_stick_button_gt_nsrts(env_name)
-    elif env_name == "doors":
+    if env_name == "doors":
         nsrts = _get_doors_gt_nsrts(env_name)
     elif env_name == "narrow_passage":
         nsrts = _get_narrow_passage_gt_nsrts(env_name)
@@ -173,167 +171,6 @@ def _get_options_by_names(env_name: str,
     options = get_gt_options(env_name)
     name_to_option = {o.name: o for o in options}
     return [name_to_option[name] for name in names]
-
-
-def _get_stick_button_gt_nsrts(env_name: str) -> Set[NSRT]:
-    """Create ground truth NSRTs for StickButtonEnv."""
-    robot_type, button_type, stick_type = _get_types_by_names(
-        env_name, ["robot", "button", "stick"])
-    Pressed, RobotAboveButton, StickAboveButton, \
-        Grasped, HandEmpty, AboveNoButton = _get_predicates_by_names(
-            env_name, ["Pressed", "RobotAboveButton",
-            "StickAboveButton", "Grasped", "HandEmpty", "AboveNoButton"])
-    RobotPressButton, PickStick, StickPressButton = _get_options_by_names(
-        env_name, ["RobotPressButton", "PickStick", "StickPressButton"])
-
-    nsrts = set()
-
-    # RobotPressButtonFromNothing
-    robot = Variable("?robot", robot_type)
-    button = Variable("?button", button_type)
-    parameters = [robot, button]
-    option_vars = [robot, button]
-    option = RobotPressButton
-    preconditions = {
-        LiftedAtom(HandEmpty, [robot]),
-        LiftedAtom(AboveNoButton, []),
-    }
-    add_effects = {
-        LiftedAtom(Pressed, [button]),
-        LiftedAtom(RobotAboveButton, [robot, button])
-    }
-    delete_effects = {LiftedAtom(AboveNoButton, [])}
-    ignore_effects: Set[Predicate] = set()
-    robot_press_button_nsrt = NSRT("RobotPressButtonFromNothing", parameters,
-                                   preconditions, add_effects, delete_effects,
-                                   ignore_effects, option, option_vars,
-                                   null_sampler)
-    nsrts.add(robot_press_button_nsrt)
-
-    # RobotPressButtonFromButton
-    robot = Variable("?robot", robot_type)
-    button = Variable("?button", button_type)
-    from_button = Variable("?from-button", button_type)
-    parameters = [robot, button, from_button]
-    option_vars = [robot, button]
-    option = RobotPressButton
-    preconditions = {
-        LiftedAtom(HandEmpty, [robot]),
-        LiftedAtom(RobotAboveButton, [robot, from_button]),
-    }
-    add_effects = {
-        LiftedAtom(Pressed, [button]),
-        LiftedAtom(RobotAboveButton, [robot, button])
-    }
-    delete_effects = {LiftedAtom(RobotAboveButton, [robot, from_button])}
-    ignore_effects = set()
-    robot_press_button_nsrt = NSRT("RobotPressButtonFromButton", parameters,
-                                   preconditions, add_effects, delete_effects,
-                                   ignore_effects, option, option_vars,
-                                   null_sampler)
-    nsrts.add(robot_press_button_nsrt)
-
-    # PickStickFromNothing
-    robot = Variable("?robot", robot_type)
-    stick = Variable("?stick", stick_type)
-    parameters = [robot, stick]
-    option_vars = [robot, stick]
-    option = PickStick
-    preconditions = {
-        LiftedAtom(HandEmpty, [robot]),
-        LiftedAtom(AboveNoButton, []),
-    }
-    add_effects = {
-        LiftedAtom(Grasped, [robot, stick]),
-    }
-    delete_effects = {LiftedAtom(HandEmpty, [robot])}
-    ignore_effects = set()
-
-    def pick_stick_sampler(state: State, goal: Set[GroundAtom],
-                           rng: np.random.Generator,
-                           objs: Sequence[Object]) -> Array:
-        del state, goal, objs  # unused
-        # Normalized x position along the long dimension of the stick, in the
-        # center of the short dimension.
-        pick_pos = rng.uniform(0, 1)
-        return np.array([pick_pos], dtype=np.float32)
-
-    pick_stick_nsrt = NSRT("PickStickFromNothing", parameters, preconditions,
-                           add_effects, delete_effects, ignore_effects, option,
-                           option_vars, pick_stick_sampler)
-    nsrts.add(pick_stick_nsrt)
-
-    # PickStickFromButton
-    robot = Variable("?robot", robot_type)
-    stick = Variable("?stick", stick_type)
-    button = Variable("?from-button", button_type)
-    parameters = [robot, stick, button]
-    option_vars = [robot, stick]
-    option = PickStick
-    preconditions = {
-        LiftedAtom(HandEmpty, [robot]),
-        LiftedAtom(RobotAboveButton, [robot, button])
-    }
-    add_effects = {
-        LiftedAtom(Grasped, [robot, stick]),
-        LiftedAtom(AboveNoButton, [])
-    }
-    delete_effects = {
-        LiftedAtom(HandEmpty, [robot]),
-        LiftedAtom(RobotAboveButton, [robot, button]),
-    }
-    ignore_effects = set()
-    pick_stick_nsrt = NSRT("PickStickFromButton", parameters, preconditions,
-                           add_effects, delete_effects, ignore_effects, option,
-                           option_vars, pick_stick_sampler)
-    nsrts.add(pick_stick_nsrt)
-
-    # StickPressButtonFromNothing
-    robot = Variable("?robot", robot_type)
-    stick = Variable("?stick", stick_type)
-    button = Variable("?button", button_type)
-    parameters = [robot, stick, button]
-    option_vars = [robot, stick, button]
-    option = StickPressButton
-    preconditions = {
-        LiftedAtom(Grasped, [robot, stick]),
-        LiftedAtom(AboveNoButton, []),
-    }
-    add_effects = {
-        LiftedAtom(StickAboveButton, [stick, button]),
-        LiftedAtom(Pressed, [button])
-    }
-    delete_effects = {LiftedAtom(AboveNoButton, [])}
-    ignore_effects = set()
-    stick_button_nsrt = NSRT("StickPressButtonFromNothing", parameters,
-                             preconditions, add_effects, delete_effects,
-                             ignore_effects, option, option_vars, null_sampler)
-    nsrts.add(stick_button_nsrt)
-
-    # StickPressButtonFromButton
-    robot = Variable("?robot", robot_type)
-    stick = Variable("?stick", stick_type)
-    button = Variable("?button", button_type)
-    from_button = Variable("?from-button", button_type)
-    parameters = [robot, stick, button, from_button]
-    option_vars = [robot, stick, button]
-    option = StickPressButton
-    preconditions = {
-        LiftedAtom(Grasped, [robot, stick]),
-        LiftedAtom(StickAboveButton, [stick, from_button])
-    }
-    add_effects = {
-        LiftedAtom(StickAboveButton, [stick, button]),
-        LiftedAtom(Pressed, [button])
-    }
-    delete_effects = {LiftedAtom(StickAboveButton, [stick, from_button])}
-    ignore_effects = set()
-    stick_button_nsrt = NSRT("StickPressButtonFromButton", parameters,
-                             preconditions, add_effects, delete_effects,
-                             ignore_effects, option, option_vars, null_sampler)
-    nsrts.add(stick_button_nsrt)
-
-    return nsrts
 
 
 def _get_doors_gt_nsrts(env_name: str) -> Set[NSRT]:
