+++ predicators/ground_truth_models/stick_button/__init__.py
@@ -1,0 +1,5 @@
+"""Ground-truth models for stick button environment and variants."""
+
+from .nsrts import StickButtonGroundTruthNSRTFactory
+
+__all__ = ["StickButtonGroundTruthNSRTFactory"]
